refactor: 拆分大文件，提取工具函数和常量
- workbench-store.ts: 提取 streaming-helpers.ts 和 message-factory.ts
  - 906 行 → 668 行，提取 SSE 流式处理和消息工厂函数
- ServiceConfig.tsx: 提取 CATEGORY_HINTS 到 constants/config-hints.ts
  - 673 行 → 512 行
- LprCalculatorTool.tsx: 提取工具函数和常量到 utils/lpr.ts
  - 744 行 → 628 行
